Hashicorp Vault: Move BouncyCastle dependency to core util Signed-off-by: Brendan Cronin <brendan.cronin@mercedes-benz.com>
